6_stream_processing/python/requirements.txt
anyio==4.9.0
attrs==25.3.0
Authlib==1.6.0
avro==1.12.0
cachetools==6.1.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
confluent-kafka==2.11.0
cryptography==45.0.5
fastavro==1.11.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
kafka-python==2.2.15
pycparser==2.22
pyspark==4.0.0
pyspark-client==4.0.0
pyspark-connect==4.0.0
requests==2.32.4
sniffio==1.3.1
typing_extensions==4.14.1
urllib3==2.5.0
